Code cleanup, preparing to separate.
