+++ pkg/web3protocol/mode-manual.go
@@ -27,21 +27,3 @@ func (client *Client) parseManualModeUrl(web3Url *Web3URL, urlMainParts map[stri
 
     return
 }
-
-
-// func handleManualMode(web3Url Web3URL) ([]byte, string, error) {
-//  var mimeType string
-//  ss := strings.Split(web3Url.RawPath, ".")
-//  if len(ss) > 1 {
-//      mimeType = mime.TypeByExtension("." + ss[len(ss)-1])
-//      log.Info("type: ", mimeType)
-//  }
-//  calldata := []byte(web3Url.RawPath)
-//  log.Info("calldata (manual): ", "0x"+hex.EncodeToString(calldata))
-//  addWeb3Header(w, "Calldata", "0x"+hex.EncodeToString(calldata))
-//  bs, werr := callContract(web3Url.Contract, web3Url.ChainId, calldata)
-//  if werr.HasError() {
-//      return nil, "", werr
-//  }
-//  return bs, mimeType, nil
-// }
